tests/requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
httpx>=0.23.0
docker>=6.0.0
websockets>=10.0
psycopg2-binary>=2.9.0
numpy>=1.22.0